Switch to using JSPM + System.js for code
