+++ bin/components/App.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+import Footer from "./Footer";
+import Header from "./Header";
+import Steps from "./Steps";
+
+export default function App() {
+  return (
+    <div>
+      <Header />
+      <Steps />
+      <Footer />
+    </div>
+  );
+}
