+++ bin/components/Check.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Check() {
+  return (
+    <span className="icon is-medium">
+      <i className="fa fa-check-circle-o"></i>
+    </span>
+  );
+}
