+++ bin/state.js
@@ -1,0 +1,9 @@
+import { observable } from "mobx";
+
+export default observable({
+  ...window.__INITIAL_STATE__,
+
+  title() {
+    return `${this.name} v${this.version}`;
+  }
+});
